django-pgvector-pdf/requirements.txt
Django>=4.0,<5.0
psycopg2-binary>=2.9
pgvector>=0.2.1
PyMuPDF>=1.22.5
sentence-transformers>=2.2.2
gunicorn>=21.2.0
transformers>=4.41.0
torch>=2.0.0
accelerate>=0.20.0
